ohana-fl-fedprox/requirements.txt
fire
monai>=1.3.0
nvflare>=2.3.0
monai_nvflare>=0.2.3
pandas
absl-py>=0.9.0
astunparse>=1.6.3
blinker>=1.4
brotlipy>=0.7.0
cachetools>=4.1.0
certifi>=2020.6.20
cffi>=1.14.0
chardet>=3.0.4
click>=7.1.2
cloudpickle>=1.4.1
cryptography==42.0.8
cycler>=0.10.0
cytoolz>=0.10.1
dask
decorator>=4.4.2
gast>=0.3.3
google-auth>=1.14.1
google-auth-oauthlib>=0.4.1
google-pasta>=0.2.0
grpcio>=1.27.2
h5py
idna
imageio>=2.8.0
Keras-Preprocessing>=1.1.0
kiwisolver>=1.2.0
Markdown>=3.1.1
matplotlib
mkl-fft>=1.1.0
mkl-random>=1.1.1
mkl-service>=2.3.0
networkx>=2.4
nibabel>=2.5.1
numpy>=1.17.4
oauthlib>=3.1.0
olefile>=0.46
opencv-python>=4.1.1.26
opt-einsum>=3.1.0
Pillow>=7.1.2
protobuf>=3.12.3
pyasn1>=0.4.8
pyasn1-modules>=0.2.7
pycparser>=2.20
PyJWT>=1.7.1
pyOpenSSL>=19.1.0
pyparsing>=2.4.7
PySocks>=1.7.1
python-dateutil>=2.8.1
PyWavelets>=1.1.1
PyYAML>=5.3.1
requests
requests-oauthlib>=1.3.0
rsa>=4.0
scikit-image>=0.16.2
scipy
six>=1.13.0
tensorboard>=2.2.1
tensorboard-plugin-wit>=1.6.0
tensorflow==2.15.0.post1
tensorflow-estimator>=2.2.0
termcolor>=1.1.0
toolz>=0.10.0
tornado>=6.0.4
urllib3>=1.25.9
Werkzeug>=1.0.1
wrapt>=1.12.1
tqdm>=4.61.1
nvidia-cudnn-cu11==8.9.0.131
nvidia-cuda-runtime-cu11==11.7.99
mlflow
pynvml